pyproject.toml
[project]
name = "src"
version = "0.0.1"
description = "A short description of the project."
authors = [{ name = "raghad" }]
readme = "README.md"
classifiers = ["Programming Language :: Python :: 3"]
dependencies = [
    "awscli",
    "ipykernel>=6.30.1",
    "matplotlib>=3.10.6",
    "pandas>=2.3.2",
    "pip",
    "python-dotenv",
    "ruff",
    "scikit-learn>=1.7.1",
    "seaborn>=0.13.2", # <- now you can add it here
    "imbalanced-learn",
    "streamlit>=1.49.1",
    "langchain>=0.3.27",
    "langchain-google-genai>=2.1.10",
]
requires-python = "~=3.10.0"
